✨ feat: Allow batch setting of base_url
