+++ controller/channel.go
@@ -67,13 +67,28 @@ func AddChannel(c *gin.Context) {
 	}
 	channel.CreatedTime = utils.GetTimestamp()
 	keys := strings.Split(channel.Key, "\n")
+
+	baseUrls := []string{}
+	if channel.BaseURL != nil && *channel.BaseURL != "" {
+		baseUrls = strings.Split(*channel.BaseURL, "\n")
+	}
 	channels := make([]model.Channel, 0, len(keys))
-	for _, key := range keys {
+	for index, key := range keys {
 		if key == "" {
 			continue
 		}
 		localChannel := channel
 		localChannel.Key = key
+		if index > 0 {
+			localChannel.Name = localChannel.Name + "_" + strconv.Itoa(index+1)
+		}
+
+		if len(baseUrls) > index && baseUrls[index] != "" {
+			localChannel.BaseURL = &baseUrls[index]
+		} else if len(baseUrls) > 0 {
+			localChannel.BaseURL = &baseUrls[0]
+		}
+
 		channels = append(channels, localChannel)
 	}
 	err = model.BatchInsertChannels(channels)
