+++ web/src/views/Channel/component/TableRow.js
@@ -220,12 +220,13 @@ export default function ChannelTableRow({ item, manageChannel, handleOpenModal, 
         </TableCell>
 
         <TableCell>{item.id}</TableCell>
-
         <TableCell>{item.name}</TableCell>
 
         <TableCell>
           <GroupLabel group={item.group} />
         </TableCell>
+
+        <TableCell>{item.tag && <Label key={item.tag}>{item.tag}</Label>}</TableCell>
 
         <TableCell>
           {!CHANNEL_OPTIONS[item.type] ? (
@@ -250,12 +251,15 @@ export default function ChannelTableRow({ item, manageChannel, handleOpenModal, 
             handle_action={handleResponseTime}
           />
         </TableCell>
+        {/* <TableCell>
+          
+        </TableCell> */}
         <TableCell>
+          {renderQuota(item.used_quota)}
           <Tooltip title={'点击更新余额'} placement="top" onClick={updateChannelBalance}>
             {renderBalance(item.type, itemBalance)}
           </Tooltip>
         </TableCell>
-        <TableCell>{renderQuota(item.used_quota)}</TableCell>
         <TableCell>
           <TextField
             id={`priority-${item.id}`}
@@ -374,7 +378,7 @@ export default function ChannelTableRow({ item, manageChannel, handleOpenModal, 
         ))}
       </StyledMenu>
       <TableRow>
-        <TableCell style={{ paddingBottom: 0, paddingTop: 0, textAlign: 'left' }} colSpan={12}>
+        <TableCell style={{ paddingBottom: 0, paddingTop: 0, textAlign: 'left' }} colSpan={20}>
           <Collapse in={openRow} timeout="auto" unmountOnExit>
             <Grid container spacing={1}>
               <Grid item xs={12}>
@@ -474,20 +478,50 @@ ChannelTableRow.propTypes = {
 function renderBalance(type, balance) {
   switch (type) {
     case 1: // OpenAI
-      return <span>${balance.toFixed(2)}</span>;
+      return (
+        <span>
+          <br />${balance.toFixed(2)}
+        </span>
+      );
     case 4: // CloseAI
-      return <span>¥{balance.toFixed(2)}</span>;
+      return (
+        <span>
+          <br />¥{balance.toFixed(2)}
+        </span>
+      );
     case 8: // 自定义
-      return <span>${balance.toFixed(2)}</span>;
+      return (
+        <span>
+          <br />${balance.toFixed(2)}
+        </span>
+      );
     case 5: // OpenAI-SB
-      return <span>¥{(balance / 10000).toFixed(2)}</span>;
+      return (
+        <span>
+          <br />¥{(balance / 10000).toFixed(2)}
+        </span>
+      );
     case 10: // AI Proxy
-      return <span>{renderNumber(balance)}</span>;
+      return (
+        <span>
+          <br />
+          {renderNumber(balance)}
+        </span>
+      );
     case 12: // API2GPT
-      return <span>¥{balance.toFixed(2)}</span>;
+      return (
+        <span>
+          <br />¥{balance.toFixed(2)}
+        </span>
+      );
     case 13: // AIGC2D
-      return <span>{renderNumber(balance)}</span>;
+      return (
+        <span>
+          <br />
+          {renderNumber(balance)}
+        </span>
+      );
     default:
-      return <span>不支持</span>;
+      return <span></span>;
   }
 }
